Add files via upload
Max
(1)Instead of identifying the image names in the CSS/HTML i've taken a more practical approach (for me) by adding new images to the image file and renaming them to image names used in the code.
(2)I've added a "Proofread" image to the main page as a backdrop
(3)I've added another "Proofread" design to the top of each of the 4 pages accessible from the main page.
(4)The image needs to be added to the page that shows our 3 tests results
(5)I've added each persons profile although I was unable to link the Assignment 1 website(s) via a hyperlink to a URL
(6)The "More Information" Pages button links to a new page. This needs to be updated when Lucky finishes his summary of the three tests.
(7)I added the tables and the summary to the Ideal Jobs section.
================================================
Having said all this , given my low mark for Assignment 1 everything I have done needs to be thoroughly checked by the team.
I don't want to drag you down to my level.